Item decorator for recycler view (#212) (#247)
* Added item decorator for recycler view

Instead of setting divider view visibility on/off, added a
HorizontalItemDivider to the recycler view(folder view only atm)

* Minor refactor

-Moved the divider setup to recycler view setup instead of adapter setup
method
-Fix for ignoring start offset in divider setup method, added colorInt
annotation

* Added ItemDecoration to Songs,Ablums,Artist and Playlists

Since the 4 fragments are based on the
AbsLibaryPageRecyclerViewFragment, I added the decoration code there,
since the recycler view setup is also there. Had to da a bit of
workaround for the PlaylistFragment and the SongsFragment, hence the
shouldDrawLine method. The check for adapter is there, because the
default viewtype is 0 and the item would't draw the line on the other
lists. There was also a bit of work regarding removing/readding the
divider because of the layout item switching.

* ItemDivider for song list in AlbumDetailActivity

* ItemDivider for both player song list fragments

* Removed leftover Color import

Used for testing

* Minor refactor

-Removed unused/commented out divider code in the adapters
-Added the divider to the PlaylistDetailActivity

* Fixed line drawing in landscape mode

Fixes the problem, where the second column would be ignored in landscape
viewing mode